+++ src-tauri/src/structs/hardware.rs
@@ -1,23 +1,24 @@
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
+use specta::Type;
 
-#[derive(Serialize)]
+#[derive(Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct MemoryInfo {
   pub size: String,
-  pub clock: u64,
+  pub clock: u32,
   pub clock_unit: String,
-  pub memory_count: usize,
-  pub total_slots: usize,
+  pub memory_count: u32,
+  pub total_slots: u32,
   pub memory_type: String,
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GraphicInfo {
   pub id: String,
   pub name: String,
   pub vendor_name: String,
-  pub clock: u64,
+  pub clock: u32,
   pub memory_size: String,
   pub memory_size_dedicated: String,
 }
